backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Flask-JWT-Extended==4.7.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.9
